fix: add nonce to multisig config
